requirements.txt
gradio
openai
PyPDF2
python-dotenv